+++ src/dplyr.cpp
@@ -10,21 +10,13 @@ bool has_no_class( const RObject& arg) {
 }
 
 bool hybridable( RObject arg ){
+    if( Rf_inherits(arg, "Date") || Rf_inherits(arg, "POSIXct") || Rf_inherits(arg, "difftime") ) return true ;
+
     if( arg.isObject() || arg.isS4() ) return false ;
     int type = arg.sexp_type() ;
     switch( type ){
         case INTSXP:
-            {
-                if( has_no_class(arg) || Rf_inherits(arg, "Date") || Rf_inherits(arg, "POSIXct") )
-                    return true ;
-                break ;
-            }
         case REALSXP:
-            {
-                if( has_no_class(arg) || Rf_inherits(arg, "Date") || Rf_inherits(arg, "POSIXct") || Rf_inherits(arg, "difftime") )
-                    return true ;
-                break ;
-            }
         case LGLSXP:
         case STRSXP:
         case CPLXSXP: